requirements.txt
appnope==0.1.0
asgi-redis==1.4.3
asgiref==1.1.2
attrs==17.2.0
autobahn==17.9.3
Automat==0.6.0
channels==1.1.8
constantly==15.1.0
daphne==1.3.0
decorator==4.1.2
dj-database-url==0.3.0
Django==1.11.6
django-crispy-forms==1.5.1
django-dotenv==1.3.0
django-extensions==1.5.5
djangorestframework==3.6.4
hyperlink==17.3.1
incremental==17.5.0
ipython==6.2.1
ipython-genutils==0.2.0
jedi==0.11.0
msgpack-python==0.4.8
olefile==0.44
parso==0.1.0
pexpect==4.2.1
pickleshare==0.7.4
Pillow==4.2.1
prompt-toolkit==1.0.15
psycopg2==2.6.1
ptyprocess==0.5.2
Pygments==2.2.0
python-form==0.2.1
pytz==2017.2
redis==2.10.6
simplegeneric==0.8.1
six==1.9.0
traitlets==4.3.2
Twisted==17.9.0
txaio==2.8.2
virtualenv==15.1.0
waitress==0.8.9
wcwidth==0.1.7
websocket-client==0.44.0
websockets==2.6
whitenoise==2.0.3
zope.interface==4.4.3
